// resources/js/components/contest/types.ts
export interface ContestRef {
    id: number;
    name: string;
    year: number;
}

export interface Ref {
    id: number;
    name: string;
}

export interface Paginator<T> {
    data: T[];
    current_page: number;
    first_page_url: string;
    from: number;
    last_page: number;
    last_page_url: string;
    links: { url: string | null; label: string; active: boolean }[];
    next_page_url: string | null;
    path: string;
    per_page: number;
    prev_page_url: string | null;
    to: number;
    total: number;
}